# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Sistema de Inventario",
        "description": "Documentación oficial de la API para la gestión de inventarios, áreas y bienes.",
        "contact": {
            "email": "soporte@tuempresa.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8081/api",
            "description": "Servidor Principal de la API"
        }
    ],
    "paths": {
...
        "/gestores": {
            "get": {
                "tags": [
                    "Gestores"
                ],
                "summary": "Listar todos los gestores",
                "description": "Listar Gestores\n\nMuestra una lista paginada de todos los gestores.",
                "operationId": "f483deff4b4b680a38f36887973dae95",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de gestores",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Gestores"
                ],
                "summary": "Registrar un nuevo gestor y su usuario",
                "description": "Crear Gestor\n\nCrea un nuevo gestor y automáticamente le crea una cuenta de usuario con Rol de Gestor (ID 2).",
                "operationId": "5a00eb1a38724e3204a5b1cc4f16c28d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "gestor_nombre",
                                    "gestor_apellidos",
                                    "gestor_correo",
                                    "usuario_pass"
                                ],
                                "properties": {
                                    "gestor_nombre": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "gestor_apellidos": {
                                        "type": "string",
                                        "example": "Pérez"
                                    },
                                    "gestor_correo": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan.gestor@example.com"
                                    },
                                    "usuario_pass": {
                                        "description": "Contraseña para el acceso al sistema",
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Gestor y usuario creados exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación (correo duplicado)"
                    },
                    "500": {
                        "description": "Error en la transacción"
                    }
                }
            }
        },
        "/gestores/{id}": {
            "get": {
                "tags": [
                    "Gestores"
                ],
                "summary": "Obtener detalles de un gestor",
                "description": "Ver Gestor\n\nMuestra los detalles de un gestor específico y su usuario asociado.",
                "operationId": "2b9b07006dcec79f3c877dafdb403f0a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del gestor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Gestor no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Gestores"
                ],
                "summary": "Actualizar gestor",
                "description": "Actualizar Gestor\n\nActualiza la información del gestor y sincroniza los cambios en su usuario asociado.",
                "operationId": "5c512b91e2dc676730b046e3bfc154b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "gestor_nombre",
                                    "gestor_apellidos",
                                    "gestor_correo"
                                ],
                                "properties": {
                                    "gestor_nombre": {
                                        "type": "string"
                                    },
                                    "gestor_apellidos": {
                                        "type": "string"
                                    },
                                    "gestor_correo": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "usuario_pass": {
                                        "description": "Opcional: Nueva contraseña",
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "gestor_id_usuario": {
                                        "description": "ID del usuario vinculado (para validación)",
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Gestor actualizado"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error en la transacción"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Gestores"
                ],
                "summary": "Eliminar gestor y su usuario",
                "description": "Eliminar Gestor\n\nElimina al gestor y también elimina su cuenta de usuario asociada.",
                "operationId": "4955cd67b657fe80b6da4ab1945ca6bc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Eliminado exitosamente"
                    },
                    "409": {
                        "description": "Conflicto: Tiene registros asociados"
                    },
                    "500": {
                        "description": "Error en la transacción"
                    }
                }
            }
        },
...
    },
...
}